+++ extensions/chatgpt-atlas/CHANGELOG.md
@@ -1,5 +1,9 @@
 # ChatGPT Atlas Changelog
 
+## [Fix Crash when not installed] - {PR_MERGE_DATE}
+
+- Fix crash when **ChatGPT Atlas** is not installed
+
 ## [Improve Bookmarks Search] - 2025-10-24
 
 - Improve the bookmarks search to use the frecency sorting algorithm as well as match the search text in the name or URL.